requirements.txt
# Core Framework
fastapi==0.115.12
uvicorn==0.34.3
starlette==0.46.2
pydantic==2.11.5
pydantic_core==2.33.2

# Data Processing
pandas==2.2.3
numpy==2.2.5
openpyxl==3.1.5

# PDF Processing
PyMuPDF==1.25.5
camelot-py==1.0.0
pypdf==5.5.0
pdfminer.six==20250506

# Visualization
dash==3.0.4
dash-bootstrap-components==2.0.3
plotly==6.0.1

# Image Processing
opencv-python==4.11.0.86
pillow






# Utilities
python-dotenv==1.1.0
python-multipart==0.0.20
tqdm==4.67.1
watchdog==6.0.0
tabulate==0.9.0
python-dateutil==2.9.0.post0
pytz==2025.2
faker
